support hudi reader (#64726)
