+++ be/test/format_v2/table_reader_test.cpp
@@ -48,20 +48,14 @@
 #include "core/data_type/data_type_number.h"
 #include "core/data_type/data_type_string.h"
 #include "core/data_type/data_type_struct.h"
-#include "exec/common/endian.h"
 #include "exprs/runtime_filter_expr.h"
 #include "exprs/vectorized_fn_call.h"
 #include "exprs/vexpr.h"
 #include "exprs/vliteral.h"
 #include "exprs/vslot_ref.h"
-#include "format/format_common.h"
-#include "format/table/deletion_vector_reader.h"
-#include "format_v2/table/iceberg_reader.h"
-#include "format_v2/table/paimon_reader.h"
 #include "gen_cpp/Exprs_types.h"
 #include "gen_cpp/PlanNodes_types.h"
 #include "io/io_common.h"
-#include "roaring/roaring64map.hh"
 #include "runtime/runtime_profile.h"
 #include "runtime/runtime_state.h"
 #include "storage/predicate/predicate_creator.h"
@@ -69,10 +63,6 @@
 
 namespace doris::format {
 namespace {
-
-LocalColumnIndex field_projection(int32_t column_id) {
-    return LocalColumnIndex {.index = column_id};
-}
 
 std::vector<int32_t> projection_ids(const std::vector<LocalColumnIndex>& projections) {
     std::vector<int32_t> ids;
@@ -209,11 +199,6 @@ VExprSPtr table_int32_literal(int32_t value) {
                                    Field::create_field<TYPE_INT>(value));
 }
 
-VExprSPtr table_int64_literal(int64_t value) {
-    return VLiteral::create_shared(std::make_shared<DataTypeInt64>(),
-                                   Field::create_field<TYPE_BIGINT>(value));
-}
-
 TExprNode table_function_node(const std::string& function_name, const DataTypePtr& return_type,
                               const std::vector<DataTypePtr>& arg_types,
                               TExprNodeType::type node_type,
@@ -280,21 +265,6 @@ VExprSPtr runtime_filter_wrapper_expr(VExprSPtr impl) {
     return RuntimeFilterExpr::create_shared(node, std::move(impl), 0, false, /*filter_id=*/1);
 }
 
-VExprSPtr table_nullable_int64_binary_predicate(const std::string& function_name,
-                                                TExprOpcode::type opcode, int slot_id,
-                                                int column_id, const std::string& column_name,
-                                                int64_t value) {
-    const auto int64_type = std::make_shared<DataTypeInt64>();
-    const auto nullable_int64_type = make_nullable(int64_type);
-    auto expr = table_function_expr(function_name, make_nullable(std::make_shared<DataTypeUInt8>()),
-                                    {nullable_int64_type, int64_type}, TExprNodeType::BINARY_PRED,
-                                    opcode);
-    expr->add_child(
-            VSlotRef::create_shared(slot_id, column_id, slot_id, nullable_int64_type, column_name));
-    expr->add_child(table_int64_literal(value));
-    return expr;
-}
-
 class NullableArrayBigintDefaultExpr final : public VExpr {
 public:
     explicit NullableArrayBigintDefaultExpr(DataTypePtr data_type)
@@ -324,67 +294,6 @@ public:
 
 private:
     std::string _name;
-};
-
-class IcebergTableReaderDeleteFileTestHelper final
-        : public doris::format::iceberg::IcebergTableReader {
-public:
-    Status parse_deletion_vector_file(const TTableFormatFileDesc& t_desc, DeleteFileDesc* desc,
-                                      bool* has_delete_file) {
-        return _parse_deletion_vector_file(t_desc, desc, has_delete_file);
-    }
-};
-
-class IcebergTableReaderScanRequestTestHelper final
-        : public doris::format::iceberg::IcebergTableReader {
-public:
-    Status init_for_scan_request_test(std::vector<ColumnDefinition> projected_columns) {
-        _query_options = std::make_unique<TQueryOptions>();
-        _query_globals = std::make_unique<TQueryGlobals>();
-        _state = std::make_unique<RuntimeState>(*_query_options, *_query_globals);
-        RETURN_IF_ERROR(init({
-                .projected_columns = std::move(projected_columns),
-                .column_predicates = {},
-                .conjuncts = {},
-                .format = FileFormat::PARQUET,
-                .scan_params = nullptr,
-                .io_ctx = nullptr,
-                .runtime_state = _state.get(),
-                .scanner_profile = nullptr,
-        }));
-
-        SplitReadOptions split_options;
-        split_options.current_range.__set_path("scan-request-test.parquet");
-        TTableFormatFileDesc table_format_params;
-        TIcebergFileDesc iceberg_params;
-        iceberg_params.__set_first_row_id(1000);
-        table_format_params.__set_iceberg_params(iceberg_params);
-        split_options.current_range.__set_table_format_params(table_format_params);
-        RETURN_IF_ERROR(prepare_split(split_options));
-
-        _delete_rows_storage = {1};
-        _delete_rows = &_delete_rows_storage;
-        return Status::OK();
-    }
-
-    Status customize_request(FileScanRequest* request) {
-        return customize_file_scan_request(request);
-    }
-
-private:
-    std::unique_ptr<TQueryOptions> _query_options;
-    std::unique_ptr<TQueryGlobals> _query_globals;
-    std::unique_ptr<RuntimeState> _state;
-    DeleteRows _delete_rows_storage;
-};
-
-class IcebergTableReaderMappingModeTestHelper final
-        : public doris::format::iceberg::IcebergTableReader {
-public:
-    TableColumnMappingMode mapping_mode_for_schema(std::vector<ColumnDefinition> file_schema) {
-        _data_reader.file_schema = std::move(file_schema);
-        return mapping_mode();
-    }
 };
 
 class TableReaderMaterializeTestHelper final : public TableReader {
@@ -525,27 +434,6 @@ std::shared_ptr<arrow::Array> build_int32_array(const std::vector<int32_t>& valu
     return finish_array(&builder);
 }
 
-std::shared_ptr<arrow::Array> build_int64_array(const std::vector<int64_t>& values) {
-    arrow::Int64Builder builder;
-    for (const auto value : values) {
-        EXPECT_TRUE(builder.Append(value).ok());
-    }
-    return finish_array(&builder);
-}
-
-std::shared_ptr<arrow::Array> build_nullable_int64_array(
-        const std::vector<std::optional<int64_t>>& values) {
-    arrow::Int64Builder builder;
-    for (const auto& value : values) {
-        if (value.has_value()) {
-            EXPECT_TRUE(builder.Append(*value).ok());
-        } else {
-            EXPECT_TRUE(builder.AppendNull().ok());
-        }
-    }
-    return finish_array(&builder);
-}
-
 std::shared_ptr<arrow::Array> build_string_array(const std::vector<std::string>& values) {
     arrow::StringBuilder builder;
     for (const auto& value : values) {
@@ -560,48 +448,6 @@ void write_parquet_file(const std::string& file_path, int32_t id, const std::str
             arrow::field("value", arrow::utf8(), false),
     });
     auto table = arrow::Table::Make(schema, {build_int32_array({id}), build_string_array({value})});
-
-    auto file_result = arrow::io::FileOutputStream::Open(file_path);
-    ASSERT_TRUE(file_result.ok()) << file_result.status();
-    std::shared_ptr<arrow::io::FileOutputStream> out = *file_result;
-
-    ::parquet::WriterProperties::Builder builder;
-    builder.version(::parquet::ParquetVersion::PARQUET_2_6);
-    builder.data_page_version(::parquet::ParquetDataPageVersion::V2);
-    builder.compression(::parquet::Compression::UNCOMPRESSED);
-    PARQUET_THROW_NOT_OK(::parquet::arrow::WriteTable(*table, arrow::default_memory_pool(), out, 1,
-                                                      builder.build()));
-}
-
-void write_iceberg_equality_delete_parquet_file(const std::string& file_path, int32_t field_id,
-                                                int32_t value) {
-    const auto metadata =
-            arrow::key_value_metadata({"PARQUET:field_id"}, {std::to_string(field_id)});
-    auto schema = arrow::schema({
-            arrow::field("id", arrow::int32(), false)->WithMetadata(metadata),
-    });
-    auto table = arrow::Table::Make(schema, {build_int32_array({value})});
-
-    auto file_result = arrow::io::FileOutputStream::Open(file_path);
-    ASSERT_TRUE(file_result.ok()) << file_result.status();
-    std::shared_ptr<arrow::io::FileOutputStream> out = *file_result;
-
-    ::parquet::WriterProperties::Builder builder;
-    builder.version(::parquet::ParquetVersion::PARQUET_2_6);
-    builder.data_page_version(::parquet::ParquetDataPageVersion::V2);
-    builder.compression(::parquet::Compression::UNCOMPRESSED);
-    PARQUET_THROW_NOT_OK(::parquet::arrow::WriteTable(*table, arrow::default_memory_pool(), out, 1,
-                                                      builder.build()));
-}
-
-void write_iceberg_equality_delete_bigint_parquet_file(const std::string& file_path,
-                                                       int32_t field_id, int64_t value) {
-    const auto metadata =
-            arrow::key_value_metadata({"PARQUET:field_id"}, {std::to_string(field_id)});
-    auto schema = arrow::schema({
-            arrow::field("id", arrow::int64(), false)->WithMetadata(metadata),
-    });
-    auto table = arrow::Table::Make(schema, {build_int64_array({value})});
 
     auto file_result = arrow::io::FileOutputStream::Open(file_path);
     ASSERT_TRUE(file_result.ok()) << file_result.status();
@@ -842,155 +688,12 @@ void write_int_pair_parquet_file(const std::string& file_path, const std::vector
                                                       write_row_group_size, builder.build()));
 }
 
-void write_iceberg_row_lineage_parquet_file(
-        const std::string& file_path, const std::vector<int32_t>& ids,
-        const std::vector<std::optional<int64_t>>& row_ids,
-        const std::vector<std::optional<int64_t>>& last_updated_sequence_numbers = {}) {
-    ASSERT_EQ(ids.size(), row_ids.size());
-    if (!last_updated_sequence_numbers.empty()) {
-        ASSERT_EQ(ids.size(), last_updated_sequence_numbers.size());
-    }
-    const auto id_metadata = arrow::key_value_metadata({"PARQUET:field_id"}, {"0"});
-    const auto row_id_metadata = arrow::key_value_metadata({"PARQUET:field_id"}, {"2147483540"});
-    const auto last_updated_sequence_number_metadata =
-            arrow::key_value_metadata({"PARQUET:field_id"}, {"2147483539"});
-    auto schema = arrow::schema({
-            arrow::field("id", arrow::int32(), false)->WithMetadata(id_metadata),
-            arrow::field("_row_id", arrow::int64(), true)->WithMetadata(row_id_metadata),
-    });
-    std::vector<std::shared_ptr<arrow::Array>> arrays = {
-            build_int32_array(ids),
-            build_nullable_int64_array(row_ids),
-    };
-    if (!last_updated_sequence_numbers.empty()) {
-        schema =
-                schema->AddField(schema->num_fields(),
-                                 arrow::field("_last_updated_sequence_number", arrow::int64(), true)
-                                         ->WithMetadata(last_updated_sequence_number_metadata))
-                        .ValueOrDie();
-        arrays.push_back(build_nullable_int64_array(last_updated_sequence_numbers));
-    }
-    auto table = arrow::Table::Make(schema, arrays);
-
-    auto file_result = arrow::io::FileOutputStream::Open(file_path);
-    ASSERT_TRUE(file_result.ok()) << file_result.status();
-    std::shared_ptr<arrow::io::FileOutputStream> out = *file_result;
-
-    ::parquet::WriterProperties::Builder builder;
-    builder.version(::parquet::ParquetVersion::PARQUET_2_6);
-    builder.data_page_version(::parquet::ParquetDataPageVersion::V2);
-    builder.compression(::parquet::Compression::UNCOMPRESSED);
-    PARQUET_THROW_NOT_OK(::parquet::arrow::WriteTable(*table, arrow::default_memory_pool(), out,
-                                                      static_cast<int64_t>(ids.size()),
-                                                      builder.build()));
-}
-
-void write_position_delete_parquet_file(const std::string& file_path,
-                                        const std::vector<std::string>& data_file_paths,
-                                        const std::vector<int64_t>& positions) {
-    auto schema = arrow::schema({
-            arrow::field("file_path", arrow::utf8(), false),
-            arrow::field("pos", arrow::int64(), false),
-    });
-    auto table = arrow::Table::Make(
-            schema, {build_string_array(data_file_paths), build_int64_array(positions)});
-
-    auto file_result = arrow::io::FileOutputStream::Open(file_path);
-    ASSERT_TRUE(file_result.ok()) << file_result.status();
-    std::shared_ptr<arrow::io::FileOutputStream> out = *file_result;
-
-    ::parquet::WriterProperties::Builder builder;
-    builder.version(::parquet::ParquetVersion::PARQUET_2_6);
-    builder.data_page_version(::parquet::ParquetDataPageVersion::V2);
-    builder.compression(::parquet::Compression::UNCOMPRESSED);
-    PARQUET_THROW_NOT_OK(::parquet::arrow::WriteTable(*table, arrow::default_memory_pool(), out,
-                                                      static_cast<int64_t>(positions.size()),
-                                                      builder.build()));
-}
-
-int64_t write_iceberg_deletion_vector_file(const std::string& file_path,
-                                           const std::vector<uint64_t>& deleted_positions) {
-    roaring::Roaring64Map rows;
-    for (const auto position : deleted_positions) {
-        rows.add(position);
-    }
-
-    const size_t bitmap_size = rows.getSizeInBytes();
-    std::vector<char> blob(4 + 4 + bitmap_size + 4);
-    rows.write(blob.data() + 8);
-
-    const uint32_t total_length = static_cast<uint32_t>(4 + bitmap_size);
-    BigEndian::Store32(blob.data(), total_length);
-    constexpr char DV_MAGIC[] = {'\xD1', '\xD3', '\x39', '\x64'};
-    memcpy(blob.data() + 4, DV_MAGIC, 4);
-    BigEndian::Store32(blob.data() + 8 + bitmap_size, 0);
-
-    std::ofstream output(file_path, std::ios::binary);
-    EXPECT_TRUE(output.is_open());
-    output.write(blob.data(), static_cast<std::streamsize>(blob.size()));
-    EXPECT_TRUE(output.good());
-    return static_cast<int64_t>(blob.size());
-}
-
-int64_t write_paimon_deletion_vector_file(const std::string& file_path,
-                                          const std::vector<uint32_t>& deleted_positions) {
-    roaring::Roaring rows;
-    for (const auto position : deleted_positions) {
-        rows.add(position);
-    }
-
-    const size_t bitmap_size = rows.getSizeInBytes();
-    const uint32_t total_length = static_cast<uint32_t>(4 + bitmap_size);
-    std::vector<char> blob(4 + total_length);
-    BigEndian::Store32(blob.data(), total_length);
-    constexpr char PAIMON_BITMAP_MAGIC[] = {'\x5E', '\x43', '\xF2', '\xD0'};
-    memcpy(blob.data() + 4, PAIMON_BITMAP_MAGIC, 4);
-    rows.write(blob.data() + 8);
-
-    std::ofstream output(file_path, std::ios::binary);
-    EXPECT_TRUE(output.is_open());
-    output.write(blob.data(), static_cast<std::streamsize>(blob.size()));
-    EXPECT_TRUE(output.good());
-    // Paimon DeletionFile.length is magic + bitmap length, excluding the leading length field.
-    return static_cast<int64_t>(total_length);
-}
-
 Block build_table_block(const std::vector<ColumnDefinition>& columns) {
     Block block;
     for (const auto& column : columns) {
         block.insert({column.type->create_column(), column.type, column.name});
     }
     return block;
-}
-
-void expect_nullable_int64_column_values(const IColumn& column,
-                                         const std::vector<int64_t>& expected_values) {
-    const auto full_column = column.convert_to_full_column_if_const();
-    const auto& nullable_column = assert_cast<const ColumnNullable&>(*full_column);
-    const auto& values =
-            assert_cast<const ColumnInt64&>(nullable_column.get_nested_column()).get_data();
-    ASSERT_EQ(nullable_column.size(), expected_values.size());
-    for (size_t row = 0; row < expected_values.size(); ++row) {
-        EXPECT_EQ(nullable_column.get_null_map_data()[row], 0);
-        EXPECT_EQ(values[row], expected_values[row]);
-    }
-}
-
-void expect_nullable_int64_column_optional_values(
-        const IColumn& column, const std::vector<std::optional<int64_t>>& expected_values) {
-    const auto full_column = column.convert_to_full_column_if_const();
-    const auto& nullable_column = assert_cast<const ColumnNullable&>(*full_column);
-    const auto& values =
-            assert_cast<const ColumnInt64&>(nullable_column.get_nested_column()).get_data();
-    ASSERT_EQ(nullable_column.size(), expected_values.size());
-    for (size_t row = 0; row < expected_values.size(); ++row) {
-        if (expected_values[row].has_value()) {
-            EXPECT_EQ(nullable_column.get_null_map_data()[row], 0);
-            EXPECT_EQ(values[row], *expected_values[row]);
-        } else {
-            EXPECT_EQ(nullable_column.get_null_map_data()[row], 1);
-        }
-    }
 }
 
 const IColumn& expect_not_null_nullable_nested_column(const IColumn& column) {
@@ -1018,50 +721,6 @@ const IColumn& expect_not_null_table_column(const Block& block, size_t position)
 
 ColumnDefinition make_table_column(int32_t id, const std::string& name, const DataTypePtr& type);
 
-DataTypePtr make_iceberg_rowid_type() {
-    return make_nullable(std::make_shared<DataTypeStruct>(
-            DataTypes {std::make_shared<DataTypeString>(), std::make_shared<DataTypeInt64>(),
-                       std::make_shared<DataTypeInt32>(), std::make_shared<DataTypeString>()},
-            Strings {"file_path", "row_pos", "partition_spec_id", "partition_data_json"}));
-}
-
-ColumnDefinition make_iceberg_row_lineage_row_id_column() {
-    return make_table_column(2147483540, "_row_id",
-                             make_nullable(std::make_shared<DataTypeInt64>()));
-}
-
-ColumnDefinition make_iceberg_last_updated_sequence_number_column() {
-    return make_table_column(2147483539, "_last_updated_sequence_number",
-                             make_nullable(std::make_shared<DataTypeInt64>()));
-}
-
-void expect_iceberg_rowid_column_values(const IColumn& column, const std::string& file_path,
-                                        const std::vector<int64_t>& row_positions,
-                                        int32_t partition_spec_id,
-                                        const std::string& partition_data_json) {
-    const auto full_column = column.convert_to_full_column_if_const();
-    const auto& nullable_column = assert_cast<const ColumnNullable&>(*full_column);
-    const auto& struct_column =
-            assert_cast<const ColumnStruct&>(nullable_column.get_nested_column());
-    const auto& file_path_column = assert_cast<const ColumnString&>(
-            expect_not_null_nullable_nested_column(struct_column.get_column(0)));
-    const auto& row_pos_column = assert_cast<const ColumnInt64&>(
-            expect_not_null_nullable_nested_column(struct_column.get_column(1)));
-    const auto& spec_id_column = assert_cast<const ColumnInt32&>(
-            expect_not_null_nullable_nested_column(struct_column.get_column(2)));
-    const auto& partition_data_column = assert_cast<const ColumnString&>(
-            expect_not_null_nullable_nested_column(struct_column.get_column(3)));
-
-    ASSERT_EQ(nullable_column.size(), row_positions.size());
-    for (size_t row = 0; row < row_positions.size(); ++row) {
-        EXPECT_EQ(nullable_column.get_null_map_data()[row], 0);
-        EXPECT_EQ(file_path_column.get_data_at(row).to_string(), file_path);
-        EXPECT_EQ(row_pos_column.get_element(row), row_positions[row]);
-        EXPECT_EQ(spec_id_column.get_element(row), partition_spec_id);
-        EXPECT_EQ(partition_data_column.get_data_at(row).to_string(), partition_data_json);
-    }
-}
-
 void expect_int32_column_values(const IColumn& column,
                                 const std::vector<int32_t>& expected_values) {
     const auto full_column = column.convert_to_full_column_if_const();
@@ -1085,120 +744,6 @@ void set_table_level_row_count(SplitReadOptions* split_options, int64_t row_coun
     split_options->current_range.__isset.table_format_params = true;
     split_options->current_range.table_format_params.__isset.table_level_row_count = true;
     split_options->current_range.table_format_params.table_level_row_count = row_count;
-}
-
-void set_iceberg_row_lineage_params(SplitReadOptions* split_options, int64_t first_row_id,
-                                    int64_t last_updated_sequence_number) {
-    TTableFormatFileDesc table_format_params;
-    TIcebergFileDesc iceberg_params;
-    iceberg_params.__set_first_row_id(first_row_id);
-    iceberg_params.__set_last_updated_sequence_number(last_updated_sequence_number);
-    table_format_params.__set_iceberg_params(iceberg_params);
-    split_options->current_range.__set_table_format_params(table_format_params);
-}
-
-void set_iceberg_rowid_params(SplitReadOptions* split_options,
-                              const std::string& original_file_path, int32_t partition_spec_id,
-                              const std::string& partition_data_json) {
-    TTableFormatFileDesc table_format_params;
-    TIcebergFileDesc iceberg_params;
-    iceberg_params.__set_original_file_path(original_file_path);
-    iceberg_params.__set_partition_spec_id(partition_spec_id);
-    iceberg_params.__set_partition_data_json(partition_data_json);
-    table_format_params.__set_iceberg_params(iceberg_params);
-    split_options->current_range.__set_table_format_params(table_format_params);
-}
-
-TIcebergDeleteFileDesc make_iceberg_deletion_vector(const std::string& path, int64_t offset,
-                                                    int64_t size) {
-    TIcebergDeleteFileDesc delete_file;
-    delete_file.__set_content(3);
-    delete_file.__set_path(path);
-    delete_file.__set_content_offset(offset);
-    delete_file.__set_content_size_in_bytes(size);
-    return delete_file;
-}
-
-TIcebergDeleteFileDesc make_iceberg_position_delete_file(const std::string& path) {
-    TIcebergDeleteFileDesc delete_file;
-    delete_file.__set_content(1);
-    delete_file.__set_path(path);
-    delete_file.__set_file_format(TFileFormatType::FORMAT_PARQUET);
-    return delete_file;
-}
-
-TIcebergDeleteFileDesc make_iceberg_equality_delete_file(const std::string& path,
-                                                         const std::vector<int32_t>& field_ids) {
-    TIcebergDeleteFileDesc delete_file;
-    delete_file.__set_content(2);
-    delete_file.__set_path(path);
-    delete_file.__set_field_ids(field_ids);
-    delete_file.__set_file_format(TFileFormatType::FORMAT_PARQUET);
-    return delete_file;
-}
-
-TFileScanRangeParams make_local_parquet_scan_params() {
-    TFileScanRangeParams scan_params;
-    scan_params.__set_file_type(TFileType::FILE_LOCAL);
-    scan_params.__set_format_type(TFileFormatType::FORMAT_PARQUET);
-    return scan_params;
-}
-
-std::shared_ptr<io::IOContext> make_io_context(io::FileReaderStats* file_reader_stats,
-                                               io::FileCacheStatistics* file_cache_stats) {
-    auto io_ctx = std::make_shared<io::IOContext>();
-    io_ctx->file_reader_stats = file_reader_stats;
-    io_ctx->file_cache_stats = file_cache_stats;
-    return io_ctx;
-}
-
-TTableFormatFileDesc make_iceberg_table_format_desc(
-        const std::string& data_file_path,
-        const std::vector<TIcebergDeleteFileDesc>& delete_files) {
-    TTableFormatFileDesc table_format_params;
-    TIcebergFileDesc iceberg_params;
-    iceberg_params.__set_format_version(2);
-    iceberg_params.__set_original_file_path(data_file_path);
-    iceberg_params.__set_delete_files(delete_files);
-    table_format_params.__set_iceberg_params(iceberg_params);
-    return table_format_params;
-}
-
-TTableFormatFileDesc make_paimon_table_format_desc(const std::string& deletion_file_path,
-                                                   int64_t offset, int64_t length) {
-    TTableFormatFileDesc table_format_params;
-    TPaimonFileDesc paimon_params;
-    paimon_params.__set_file_format("parquet");
-    TPaimonDeletionFileDesc deletion_file;
-    deletion_file.__set_path(deletion_file_path);
-    deletion_file.__set_offset(offset);
-    deletion_file.__set_length(length);
-    paimon_params.__set_deletion_file(deletion_file);
-    table_format_params.__set_paimon_params(paimon_params);
-    return table_format_params;
-}
-
-std::vector<int32_t> read_iceberg_ids(doris::format::iceberg::IcebergTableReader* reader,
-                                      const std::vector<ColumnDefinition>& projected_columns) {
-    std::vector<int32_t> ids;
-    bool eos = false;
-    while (!eos) {
-        Block block = build_table_block(projected_columns);
-        auto status = reader->get_block(&block, &eos);
-        if (!status.ok()) {
-            ADD_FAILURE() << status;
-            return ids;
-        }
-        if (block.rows() == 0) {
-            continue;
-        }
-        const auto& id_column =
-                assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 0));
-        for (size_t row = 0; row < block.rows(); ++row) {
-            ids.push_back(id_column.get_element(row));
-        }
-    }
-    return ids;
 }
 
 int64_t parquet_column_start_offset(const ::parquet::ColumnChunkMetaData& column_metadata) {
@@ -1364,11 +909,6 @@ VExprContextSPtr prepared_conjunct(RuntimeState* state, const VExprSPtr& expr) {
     status = ctx->open(state);
     EXPECT_TRUE(status.ok()) << status;
     return ctx;
-}
-
-void apply_final_conjuncts(Block* block, const VExprContextSPtrs& conjuncts) {
-    const auto status = VExprContext::filter_block(conjuncts, block, block->columns());
-    ASSERT_TRUE(status.ok()) << status;
 }
 
 struct FakeFileReaderState {
@@ -3915,1183 +3455,6 @@ TEST(TableReaderTest, RuntimeFilterOnConstantPartitionIsNotPreExecuted) {
 
     ASSERT_TRUE(reader.close().ok());
     std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergVirtualColumnsUseRowLineageMetadata) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_virtual_columns_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_iceberg_last_updated_sequence_number_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {prepared_conjunct(
-                                            &state, table_int32_greater_than_expr(2, 2, 1))},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 1000, 77);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 2));
-
-    ASSERT_EQ(block.rows(), 2);
-    EXPECT_EQ(id_column.get_element(0), 2);
-    EXPECT_EQ(id_column.get_element(1), 3);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {1001, 1002});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {77, 77});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergRowLineageUsesPhysicalRowIdAndFillsNulls) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_physical_row_id_fill_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_iceberg_row_lineage_parquet_file(file_path, {1, 2, 3}, {7000, std::nullopt, 7002},
-                                           {80, std::nullopt, 82});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(
-            2147483540, "_row_id", make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(
-            make_table_column(2147483539, "_last_updated_sequence_number",
-                              make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 1000, 77);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    ASSERT_EQ(block.rows(), 3);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {7000, 1001, 7002});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {80, 77, 82});
-    expect_int32_column_values(*block.get_by_position(2).column, {1, 2, 3});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergPhysicalRowIdKeepsNullsWithoutFirstRowId) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_physical_row_id_no_first_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_iceberg_row_lineage_parquet_file(file_path, {1, 2, 3}, {7000, std::nullopt, 7002},
-                                           {80, std::nullopt, 82});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(
-            2147483540, "_row_id", make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(
-            make_table_column(2147483539, "_last_updated_sequence_number",
-                              make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    ASSERT_TRUE(reader.prepare_split(build_split_options(file_path)).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    ASSERT_EQ(block.rows(), 3);
-    expect_nullable_int64_column_optional_values(
-            *block.get_by_position(0).column,
-            std::vector<std::optional<int64_t>> {7000, std::nullopt, 7002});
-    expect_nullable_int64_column_optional_values(
-            *block.get_by_position(1).column,
-            std::vector<std::optional<int64_t>> {80, std::nullopt, 82});
-    expect_int32_column_values(*block.get_by_position(2).column, {1, 2, 3});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergMissingRowIdStaysNullWithoutFirstRowId) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_missing_row_id_no_first_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_iceberg_last_updated_sequence_number_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    ASSERT_TRUE(reader.prepare_split(build_split_options(file_path)).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    ASSERT_EQ(block.rows(), 3);
-    expect_nullable_int64_column_optional_values(
-            *block.get_by_position(0).column,
-            std::vector<std::optional<int64_t>> {std::nullopt, std::nullopt, std::nullopt});
-    expect_nullable_int64_column_optional_values(
-            *block.get_by_position(1).column,
-            std::vector<std::optional<int64_t>> {std::nullopt, std::nullopt, std::nullopt});
-    expect_int32_column_values(*block.get_by_position(2).column, {1, 2, 3});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergRowIdPredicateFiltersAfterRowLineageMaterialization) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_row_id_finalize_filter_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_iceberg_row_lineage_parquet_file(file_path, {1, 2, 3}, {7000, std::nullopt, 7002},
-                                           {80, std::nullopt, 82});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(
-            2147483540, "_row_id", make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(
-            make_table_column(2147483539, "_last_updated_sequence_number",
-                              make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    VExprContextSPtrs conjuncts = {prepared_conjunct(
-            &state,
-            table_nullable_int64_binary_predicate("eq", TExprOpcode::EQ, 0, 0, "_row_id", 1001))};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = conjuncts,
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 1000, 77);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 3);
-
-    apply_final_conjuncts(&block, conjuncts);
-    ASSERT_EQ(block.rows(), 1);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {1001});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {77});
-    expect_int32_column_values(*block.get_by_position(2).column, {2});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergLastUpdatedSequencePredicateFiltersAfterMaterialization) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_sequence_finalize_filter_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_iceberg_row_lineage_parquet_file(file_path, {1, 2, 3}, {7000, std::nullopt, 7002},
-                                           {80, std::nullopt, 82});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(
-            2147483540, "_row_id", make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(
-            make_table_column(2147483539, "_last_updated_sequence_number",
-                              make_nullable(std::make_shared<DataTypeInt64>())));
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    VExprContextSPtrs conjuncts = {prepared_conjunct(
-            &state, table_nullable_int64_binary_predicate("eq", TExprOpcode::EQ, 1, 1,
-                                                          "_last_updated_sequence_number", 77))};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = conjuncts,
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 1000, 77);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 3);
-
-    apply_final_conjuncts(&block, conjuncts);
-    ASSERT_EQ(block.rows(), 1);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {1001});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {77});
-    expect_int32_column_values(*block.get_by_position(2).column, {2});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergRowidVirtualColumnUsesDataFilePosition) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_rowid_virtual_column_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(
-            make_table_column(-1, BeConsts::ICEBERG_ROWID_COL, make_iceberg_rowid_type()));
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {prepared_conjunct(
-                                            &state, table_int32_greater_than_expr(1, 1, 1))},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    const auto original_file_path = "s3://bucket/table/data/original.parquet";
-    const auto partition_data_json = R"({"part":"p1"})";
-    set_iceberg_rowid_params(&split_options, original_file_path, 17, partition_data_json);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    ASSERT_EQ(block.rows(), 2);
-    expect_iceberg_rowid_column_values(*block.get_by_position(0).column, original_file_path, {1, 2},
-                                       17, partition_data_json);
-    expect_int32_column_values(*block.get_by_position(1).column, {2, 3});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergVirtualColumnsKeepRowLineageAfterConjunctFiltering) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_virtual_columns_conjunct_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_iceberg_last_updated_sequence_number_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {prepared_conjunct(
-                                            &state, table_int32_greater_than_expr(2, 2, 1))},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 3000, 88);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 2));
-
-    ASSERT_EQ(block.rows(), 2);
-    EXPECT_EQ(id_column.get_element(0), 2);
-    EXPECT_EQ(id_column.get_element(1), 3);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {3001, 3002});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {88, 88});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergVirtualColumnsKeepRowLineageAfterRowGroupPredicatePruning) {
-    const auto test_dir = std::filesystem::temp_directory_path() /
-                          "doris_iceberg_virtual_columns_row_group_predicate_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    // ColumnPredicate is used for row-group/statistics pruning. Keep one row per row group so
-    // id > 2 prunes the first two row groups and leaves only the third file-local row.
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"}, 1);
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_iceberg_last_updated_sequence_number_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    TableColumnPredicates column_predicates;
-    add_column_predicate(&column_predicates, GlobalIndex(2),
-                         create_comparison_predicate<PredicateType::GT>(
-                                 0, "id", make_nullable(std::make_shared<DataTypeInt32>()),
-                                 Field::create_field<TYPE_INT>(2), false));
-
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = std::move(column_predicates),
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    set_iceberg_row_lineage_params(&split_options, 4000, 99);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 2));
-
-    ASSERT_EQ(block.rows(), 1);
-    EXPECT_EQ(id_column.get_element(0), 3);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {4002});
-    expect_nullable_int64_column_values(*block.get_by_position(1).column, {99});
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergDeletionVectorUsesTableReaderDeleteFileInterface) {
-    TTableFormatFileDesc table_format_desc;
-    TIcebergFileDesc iceberg_desc;
-    iceberg_desc.__set_format_version(2);
-    iceberg_desc.__set_delete_files({make_iceberg_deletion_vector("dv.bin", 8, 128)});
-    table_format_desc.__set_iceberg_params(iceberg_desc);
-
-    IcebergTableReaderDeleteFileTestHelper reader;
-    DeleteFileDesc desc;
-    bool has_delete_file = false;
-    ASSERT_TRUE(reader.parse_deletion_vector_file(table_format_desc, &desc, &has_delete_file).ok());
-
-    EXPECT_TRUE(has_delete_file);
-    EXPECT_EQ(desc.path, "dv.bin");
-    EXPECT_EQ(desc.start_offset, 8);
-    EXPECT_EQ(desc.size, 128);
-    EXPECT_EQ(desc.file_size, -1);
-    EXPECT_EQ(desc.format, DeleteFileDesc::Format::ICEBERG);
-}
-
-TEST(TableReaderTest, IcebergDeletionVectorRejectsMultipleDeleteFiles) {
-    TTableFormatFileDesc table_format_desc;
-    TIcebergFileDesc iceberg_desc;
-    iceberg_desc.__set_format_version(2);
-    iceberg_desc.__set_delete_files({make_iceberg_deletion_vector("dv-a.bin", 8, 128),
-                                     make_iceberg_deletion_vector("dv-b.bin", 16, 256)});
-    table_format_desc.__set_iceberg_params(iceberg_desc);
-
-    IcebergTableReaderDeleteFileTestHelper reader;
-    DeleteFileDesc desc;
-    bool has_delete_file = false;
-    auto status = reader.parse_deletion_vector_file(table_format_desc, &desc, &has_delete_file);
-
-    EXPECT_FALSE(status.ok());
-}
-
-TEST(TableReaderTest, IcebergTableReaderAppliesDeletionVectorFile) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_deletion_vector_file_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto dv_path = (test_dir / "delete-vector.bin").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3, 4, 5}, {10, 20, 30, 40, 50},
-                                {"one", "two", "three", "four", "five"});
-    const auto dv_size = write_iceberg_deletion_vector_file(dv_path, {0, 4});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                                    .push_down_agg_type = TPushAggOp::type::COUNT,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_deletion_vector(dv_path, 0, dv_size)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({2, 3, 4}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, PaimonTableReaderAppliesBitmapDeletionVectorFile) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_paimon_deletion_vector_file_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto dv_path = (test_dir / "delete-vector.bin").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3, 4, 5}, {10, 20, 30, 40, 50},
-                                {"one", "two", "three", "four", "five"});
-    const auto dv_length = write_paimon_deletion_vector_file(dv_path, {0, 4});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::paimon::PaimonReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(
-            make_paimon_table_format_desc(dv_path, 0, dv_length));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    std::vector<int32_t> ids;
-    bool eos = false;
-    while (!eos) {
-        Block block = build_table_block(projected_columns);
-        ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-        if (block.rows() == 0) {
-            continue;
-        }
-        const auto& id_column =
-                assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 0));
-        for (size_t row = 0; row < block.rows(); ++row) {
-            ids.push_back(id_column.get_element(row));
-        }
-    }
-    EXPECT_EQ(ids, std::vector<int32_t>({2, 3, 4}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergTableReaderDoesNotPushDownAggregateWithDeletes) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_aggregate_delete_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto dv_path = (test_dir / "delete-vector.bin").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    const auto dv_size = write_iceberg_deletion_vector_file(dv_path, {0});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                                    .push_down_agg_type = TPushAggOp::type::COUNT,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_deletion_vector(dv_path, 0, dv_size)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 2);
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 0));
-    EXPECT_EQ(id_column.get_element(0), 2);
-    EXPECT_EQ(id_column.get_element(1), 3);
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-// Covers TopN lazy materialization on Iceberg schema-evolution tables. The first-phase scan adds a
-// synthesized GLOBAL_ROWID column to the file schema. That virtual column must not make Iceberg
-// fall back from field-id mapping to name mapping, otherwise renamed columns are read as defaults
-// from old files.
-TEST(TableReaderTest, IcebergMappingModeIgnoresGlobalRowIdVirtualColumn) {
-    IcebergTableReaderMappingModeTestHelper reader;
-    std::vector<ColumnDefinition> file_schema {
-            make_file_column(1, "id", std::make_shared<DataTypeInt32>()),
-            make_file_column(2, "name", std::make_shared<DataTypeString>()),
-            global_rowid_column_definition(),
-    };
-
-    EXPECT_EQ(reader.mapping_mode_for_schema(std::move(file_schema)),
-              TableColumnMappingMode::BY_FIELD_ID);
-}
-
-// Covers the fallback side of the previous case. Only synthesized columns are ignored; a real data
-// column without an Iceberg field id still disables field-id mapping.
-TEST(TableReaderTest, IcebergMappingModeRequiresFieldIdsForDataColumns) {
-    IcebergTableReaderMappingModeTestHelper reader;
-    std::vector<ColumnDefinition> file_schema {
-            make_file_column(1, "id", std::make_shared<DataTypeInt32>()),
-            make_file_column(2, "name", std::make_shared<DataTypeString>()),
-            global_rowid_column_definition(),
-    };
-    file_schema[1].identifier = Field {};
-
-    EXPECT_EQ(reader.mapping_mode_for_schema(std::move(file_schema)),
-              TableColumnMappingMode::BY_NAME);
-}
-
-TEST(TableReaderTest, IcebergTableReaderDoesNotPushDownAggregateWithPositionDelete) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_aggregate_position_delete_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_position_delete_parquet_file(delete_file_path, {file_path}, {1});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                                    .push_down_agg_type = TPushAggOp::type::COUNT,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_position_delete_file(delete_file_path)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 2);
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 0));
-    EXPECT_EQ(id_column.get_element(0), 1);
-    EXPECT_EQ(id_column.get_element(1), 3);
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergTableLevelCountUsesAssignedRowCountWithPositionDelete) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_table_level_count_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_position_delete_parquet_file(delete_file_path, {file_path}, {1});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    TQueryOptions query_options;
-    query_options.__set_batch_size(10);
-    RuntimeState state {query_options, TQueryGlobals()};
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = nullptr,
-                                    .io_ctx = nullptr,
-                                    .runtime_state = &state,
-                                    .scanner_profile = nullptr,
-                                    .push_down_agg_type = TPushAggOp::type::COUNT,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_position_delete_file(delete_file_path)}));
-    set_table_level_row_count(&split_options, 5);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    EXPECT_EQ(block.rows(), 5);
-
-    block = build_table_block(projected_columns);
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    EXPECT_TRUE(eos);
-    EXPECT_EQ(block.rows(), 0);
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergPositionDeleteFallsBackToSplitPath) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_position_delete_path_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_position_delete_parquet_file(delete_file_path, {file_path}, {1});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    TTableFormatFileDesc table_format_params;
-    TIcebergFileDesc iceberg_params;
-    iceberg_params.__set_format_version(2);
-    iceberg_params.__set_delete_files({make_iceberg_position_delete_file(delete_file_path)});
-    table_format_params.__set_iceberg_params(iceberg_params);
-    split_options.current_range.__set_table_format_params(table_format_params);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({1, 3}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergTableReaderDoesNotPushDownAggregateWithEqualityDelete) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_aggregate_equality_delete_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "equality-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_iceberg_equality_delete_parquet_file(delete_file_path, 0, 2);
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                                    .push_down_agg_type = TPushAggOp::type::COUNT,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_equality_delete_file(delete_file_path, {0})}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 2);
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 0));
-    EXPECT_EQ(id_column.get_element(0), 1);
-    EXPECT_EQ(id_column.get_element(1), 3);
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergEqualityDeleteCastsDataColumnToDeleteKeyType) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_equality_delete_cast_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "equality-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_iceberg_equality_delete_bigint_parquet_file(delete_file_path, 0, 2);
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_equality_delete_file(delete_file_path, {0})}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({1, 3}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergPositionDeleteOnlyMatchesOriginalDataFilePath) {
-    const auto test_dir = std::filesystem::temp_directory_path() /
-                          "doris_iceberg_position_delete_path_match_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto other_file_path = (test_dir / "other.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_position_delete_parquet_file(delete_file_path, {other_file_path, file_path}, {0, 1});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_position_delete_file(delete_file_path)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({1, 3}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergRowLineageRemainsFileLocalAfterDeleteFiltering) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_row_lineage_delete_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3}, {10, 20, 30}, {"one", "two", "three"});
-    write_position_delete_parquet_file(delete_file_path, {file_path}, {1});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    TTableFormatFileDesc table_format_params = make_iceberg_table_format_desc(
-            file_path, {make_iceberg_position_delete_file(delete_file_path)});
-    table_format_params.iceberg_params.__set_first_row_id(1000);
-    split_options.current_range.__set_table_format_params(table_format_params);
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    Block block = build_table_block(projected_columns);
-    bool eos = false;
-    ASSERT_TRUE(reader.get_block(&block, &eos).ok());
-    ASSERT_FALSE(eos);
-    ASSERT_EQ(block.rows(), 2);
-    expect_nullable_int64_column_values(*block.get_by_position(0).column, {1000, 1002});
-    const auto& id_column = assert_cast<const ColumnInt32&>(expect_not_null_table_column(block, 1));
-    EXPECT_EQ(id_column.get_element(0), 1);
-    EXPECT_EQ(id_column.get_element(1), 3);
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergTableReaderAppliesPositionDeleteFile) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_position_delete_file_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto delete_file_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3, 4, 5}, {10, 20, 30, 40, 50},
-                                {"one", "two", "three", "four", "five"});
-    write_position_delete_parquet_file(delete_file_path, {file_path, file_path}, {1, 3});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_position_delete_file(delete_file_path)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({1, 3, 5}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, IcebergTableReaderMergesDeletionVectorAndPositionDeleteFiles) {
-    const auto test_dir =
-            std::filesystem::temp_directory_path() / "doris_iceberg_delete_files_merge_test";
-    std::filesystem::remove_all(test_dir);
-    std::filesystem::create_directories(test_dir);
-
-    const auto file_path = (test_dir / "split.parquet").string();
-    const auto dv_path = (test_dir / "delete-vector.bin").string();
-    const auto position_delete_path = (test_dir / "position-delete.parquet").string();
-    write_int_pair_parquet_file(file_path, {1, 2, 3, 4, 5}, {10, 20, 30, 40, 50},
-                                {"one", "two", "three", "four", "five"});
-    const auto dv_size = write_iceberg_deletion_vector_file(dv_path, {0});
-    write_position_delete_parquet_file(position_delete_path, {file_path, file_path}, {3, 3});
-
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    RuntimeProfile profile("test_profile");
-    RuntimeState state {TQueryOptions(), TQueryGlobals()};
-    auto scan_params = make_local_parquet_scan_params();
-    io::FileReaderStats file_reader_stats;
-    io::FileCacheStatistics file_cache_stats;
-    auto io_ctx = make_io_context(&file_reader_stats, &file_cache_stats);
-    ShardedKVCache cache(1);
-    doris::format::iceberg::IcebergTableReader reader;
-    ASSERT_TRUE(reader.init({
-                                    .projected_columns = projected_columns,
-                                    .column_predicates = {},
-                                    .conjuncts = {},
-                                    .format = FileFormat::PARQUET,
-                                    .scan_params = &scan_params,
-                                    .io_ctx = io_ctx,
-                                    .runtime_state = &state,
-                                    .scanner_profile = &profile,
-                            })
-                        .ok());
-
-    auto split_options = build_split_options(file_path);
-    split_options.cache = &cache;
-    split_options.current_range.__set_table_format_params(make_iceberg_table_format_desc(
-            file_path, {make_iceberg_deletion_vector(dv_path, 0, dv_size),
-                        make_iceberg_position_delete_file(position_delete_path)}));
-    ASSERT_TRUE(reader.prepare_split(split_options).ok());
-
-    EXPECT_EQ(read_iceberg_ids(&reader, projected_columns), std::vector<int32_t>({2, 3, 5}));
-
-    ASSERT_TRUE(reader.close().ok());
-    std::filesystem::remove_all(test_dir);
-}
-
-TEST(TableReaderTest, RowPositionDeletePredicateColumnIsNotRepeatedAsOutputColumn) {
-    const auto row_position_column_id = ROW_POSITION_COLUMN_ID;
-    std::vector<ColumnDefinition> projected_columns;
-    projected_columns.push_back(make_iceberg_row_lineage_row_id_column());
-    projected_columns.push_back(make_table_column(0, "id", std::make_shared<DataTypeInt32>()));
-
-    IcebergTableReaderScanRequestTestHelper reader;
-    ASSERT_TRUE(reader.init_for_scan_request_test(projected_columns).ok());
-
-    FileScanRequest request;
-    request.non_predicate_columns.push_back(field_projection(0));
-    request.local_positions.emplace(LocalColumnId(0), LocalIndex(0));
-
-    ASSERT_TRUE(reader.customize_request(&request).ok());
-
-    EXPECT_EQ(projection_ids(request.predicate_columns),
-              std::vector<int32_t>({row_position_column_id}));
-    EXPECT_EQ(projection_ids(request.non_predicate_columns), std::vector<int32_t>({0}));
-    ASSERT_TRUE(request.local_positions.contains(LocalColumnId(row_position_column_id)));
-    EXPECT_EQ(request.local_positions.at(LocalColumnId(row_position_column_id)).value(), 1);
-    ASSERT_TRUE(request.conjuncts.empty());
-    ASSERT_EQ(request.delete_conjuncts.size(), 1);
-    EXPECT_NE(request.delete_conjuncts[0], nullptr);
 }
 
 TEST(TableReaderTest, ParquetReaderReadsOnlyRowGroupsInFileRange) {
